Performed required changes
- Removed !important from CSS
- Replace arrow with normal functions
- Refactored if...else
- Changed equalities to triple ===
- Remove listener on switcher deactivate
- Changed DOM variables to $camelCase
- Minor changes to layerSwitcher.html
