+++ src/og/utils/functionComposition.js
@@ -1,8 +1,6 @@
 'use strict'
 
-// HELPER for function composition
-
-// Basically, it's a function used to compose other functions,
+// Used to compose other functions,
 // in a specific order to return to desired output.
 
 const compose = x => ({
@@ -13,16 +11,3 @@ const compose = x => ({
 
 export {compose}
 
-// Example
-
-// const lower = str => str.toLowerCase();
-// const sanitize = str => str.replace(/[^a-z0-9 -]/g, '');
-// const clean = str => str.replace(/\\\\s+/gm, '-');
-
-// const slugify = str => compose(str)
-//   .run(lower)
-//   .run(sanitize)
-//   .run(clean)
-//   .end();
-
-// console.log(slugify('I love $$$ noodles')); // i-love-noodles
